feat: refactor for testability and add a config bootstrapper
Returning a list of Doers from serving.setup makes Sally usable from an end-to-end simulation test, specifically from SignifyPy. This is how I diagnosed the double present bug.

And the config bootstrapper allows Sally to be started without a preexisting AID as the bootstrapper will OOBI with the configured witnesses prior to starting up. This does not matter as much any more now that direct mode is used since we tend to avoid using witnesses when we use direct mode. The config bootstrapper is useful only when the auto-incepted AID relies on a witness for inception on first startup.